backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
Pillow==10.2.0
python-decouple==3.8
